+++ examples/data-sources/aiven_cmk_accessor_aws/data-source.tf
@@ -1,0 +1,7 @@
+data "aiven_cmk_accessor_aws" "example" {
+  project = "my-project"
+
+  /* COMPUTED FIELDS
+  principal = "foo"
+  */
+}